Remove xDS support from unsecure builds
