+++ BUILD.gn
@@ -546,9 +546,7 @@ config("grpc_config") {
         "src/core/ext/xds/xds_api.h",
         "src/core/ext/xds/xds_bootstrap.cc",
         "src/core/ext/xds/xds_bootstrap.h",
-        "src/core/ext/xds/xds_channel.h",
         "src/core/ext/xds/xds_channel_args.h",
-        "src/core/ext/xds/xds_channel_secure.cc",
         "src/core/ext/xds/xds_client.cc",
         "src/core/ext/xds/xds_client.h",
         "src/core/ext/xds/xds_client_stats.cc",
