+++ config.w32
@@ -192,7 +192,6 @@ if (PHP_GRPC != "no") {
     "src\\core\\ext\\upb-generated\\validate\\validate.upb.c " +
     "src\\core\\ext\\xds\\xds_api.cc " +
     "src\\core\\ext\\xds\\xds_bootstrap.cc " +
-    "src\\core\\ext\\xds\\xds_channel_secure.cc " +
     "src\\core\\ext\\xds\\xds_client.cc " +
     "src\\core\\ext\\xds\\xds_client_stats.cc " +
     "src\\core\\lib\\avl\\avl.cc " +
