+++ gRPC-C++.podspec
@@ -370,7 +370,6 @@ Pod::Spec.new do |s|
                       'src/core/ext/upb-generated/validate/validate.upb.h',
                       'src/core/ext/xds/xds_api.h',
                       'src/core/ext/xds/xds_bootstrap.h',
-                      'src/core/ext/xds/xds_channel.h',
                       'src/core/ext/xds/xds_channel_args.h',
                       'src/core/ext/xds/xds_client.h',
                       'src/core/ext/xds/xds_client_stats.h',
@@ -870,7 +869,6 @@ Pod::Spec.new do |s|
                               'src/core/ext/upb-generated/validate/validate.upb.h',
                               'src/core/ext/xds/xds_api.h',
                               'src/core/ext/xds/xds_bootstrap.h',
-                              'src/core/ext/xds/xds_channel.h',
                               'src/core/ext/xds/xds_channel_args.h',
                               'src/core/ext/xds/xds_client.h',
                               'src/core/ext/xds/xds_client_stats.h',
