+++ gRPC-Core.podspec
@@ -532,9 +532,7 @@ Pod::Spec.new do |s|
                       'src/core/ext/xds/xds_api.h',
                       'src/core/ext/xds/xds_bootstrap.cc',
                       'src/core/ext/xds/xds_bootstrap.h',
-                      'src/core/ext/xds/xds_channel.h',
                       'src/core/ext/xds/xds_channel_args.h',
-                      'src/core/ext/xds/xds_channel_secure.cc',
                       'src/core/ext/xds/xds_client.cc',
                       'src/core/ext/xds/xds_client.h',
                       'src/core/ext/xds/xds_client_stats.cc',
@@ -1282,7 +1280,6 @@ Pod::Spec.new do |s|
                               'src/core/ext/upb-generated/validate/validate.upb.h',
                               'src/core/ext/xds/xds_api.h',
                               'src/core/ext/xds/xds_bootstrap.h',
-                              'src/core/ext/xds/xds_channel.h',
                               'src/core/ext/xds/xds_channel_args.h',
                               'src/core/ext/xds/xds_client.h',
                               'src/core/ext/xds/xds_client_stats.h',
