+++ grpc.gemspec
@@ -450,9 +450,7 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/ext/xds/xds_api.h )
   s.files += %w( src/core/ext/xds/xds_bootstrap.cc )
   s.files += %w( src/core/ext/xds/xds_bootstrap.h )
-  s.files += %w( src/core/ext/xds/xds_channel.h )
   s.files += %w( src/core/ext/xds/xds_channel_args.h )
-  s.files += %w( src/core/ext/xds/xds_channel_secure.cc )
   s.files += %w( src/core/ext/xds/xds_client.cc )
   s.files += %w( src/core/ext/xds/xds_client.h )
   s.files += %w( src/core/ext/xds/xds_client_stats.cc )
